backend/requirements.txt
# --- Web Framework & API ---
fastapi
uvicorn[standard]
python-multipart # Cần cho xử lý Form và UploadFile

# --- Data Science & Computer Vision ---
numpy<2.0  # Fix compatibility with torch and other ML libraries (NumPy 2.x breaks torch)
Pillow
opencv-python<4.11.0  # Compatible with numpy 1.x

# --- Machine Learning / Deep Learning ---
torch
torchvision
# Bạn nên cài đặt phiên bản PyTorch dựa trên môi trường GPU/CPU của VPS.
# Nếu là CPU, bạn có thể cài đặt đơn giản qua pip.
# Nếu là GPU, hãy kiểm tra trang web PyTorch để có lệnh cài đặt chính xác.

# --- Face Recognition & Biometric ---
facenet-pytorch
scikit-learn # Được sử dụng cho normalize và cosine_similarity
# Nếu dùng GPU, hãy đảm bảo các thư viện này được cài đặt sau khi có PyTorch/CUDA

# --- Database & Authentication ---
motor # AsyncIOMotorClient cho MongoDB
passlib[bcrypt] # Hashing mật khẩu

# --- Security ---
python-jose[cryptography] # JWT token handling
python-dotenv # Environment variables management

# --- Other Utilities ---
# Note: Không cần bson riêng, pymongo (dependency của motor) đã có bson built-in
# Nếu có lỗi import bson, uninstall bson package: pip uninstall bson